Created `database/engine_data.py`
