+++ website/gameEngine.py
@@ -6,13 +6,13 @@ from datetime import datetime
 import pickle
 import logging
 import time
+
+from .database.engine_data import EmissionData, WeatherData
 from . import db
 from .database.database import (
     Under_construction,
     Shipment,
     Active_facilites,
-    WeatherData,
-    EmissionData,
 )
 from website.api import ws
