+++ website/init_test_players.py
@@ -1,6 +1,8 @@
 from werkzeug.security import generate_password_hash
 from pathlib import Path
 
+from .database.engine_data import CircularBufferNetwork, CircularBufferPlayer
+
 from .database.map import Hex
 
 from .database.player import Network, Player
@@ -8,8 +10,6 @@ from .database.database import (
     Under_construction,
     Shipment,
     Resource_on_sale,
-    CircularBufferPlayer,
-    CircularBufferNetwork,
 )
 from . import db
 import pickle
